All online user display fonctionnality
I add all new user connected to the server on a ArrayList of type "User" (new class User), and I created new methods for update the textarea of online user. And I also added a method for go on the last line of a text are when a new message is posted.